Patch from Art Yerkes:
1. I removed the global keyboard pointer, and instead acquire a new copy of
the pointer for each thread that needs it.  This makes a lock unnecessary,
but is probably less efficient than it could be.

2. Keyboard DLL names are read from the registry in the same way as on win2k.
The (Default) value is taken from HKLM\System\currentcontrolset\nls\locale,
and used to form the key HKLM\System\currentcontrolset\<8-digit-locale>
And get the layout file from the "Layout File" value.  I added the entries
for kbdus in hivesys.inf.  When others make keyboard layouts, they should be
added here.