Checkpoint:
I've been steadily working on TCP for the past few weeks.  I took a step
back/forward last week when I started addressing accessing memory in the
client program from afd.sys by using a worker thread (This will mean
that our afd will work right when client memory is on evicted pages, etc).
There are a few things that I obviously don't understand perfectly yet
about using kernel threads (or I might be overlooking something stupid I
did) so I'm still going through some slow work on it to get it perfect.
This weekend is a holiday in the US, so I have some level of hope.

In addition, I realized that our accept implementation needs substantial
work to be correct, and the main decision we need to make is not
finished yet.  I don't see accept as a blocker for the next commit,
but I will keep working on it until it works correctly afterward.  It's
likely that it will appear one or two releases later than client
functionality unless someone swoops in and finishes it off.

At the tcpip.sys layer, we need to test the code that passes through the
exception, and writable flags a bit better.

At the oskittcp layer, we need to verify that each case that can cause
a socket state change calls our wakeup with appropriate flags set.  We
lack some general functionality from FreeBSD that works on generic file
descriptors that needs to be supplemented in our code.

At each layer, we need to test and complete code that cancels requests.