widl: Fix context handle "cannot be null" detection.
The "cannot be null" attibute is applied to a parameter if and only if 
it is in-only.

widl: Fix the conditions under which array freeing code is output.

It doesn't matter whether or not the array is declared as a pointer or 
is declared using array subscripts - the array is still allocated by the 
unmarshalling function and so needs to be freed.

<rob at codeweavers.com>