[VGAFONTS] Add bitmap font for code page 862. Hebrew glyphs are from X11 font "Schumacher Clean" (see license below), modified by Baruch Rutman (peterooch*at*gmail*dot*com). CORE-6912.
Copyright 1989 Dale Schumacher, dal@syntel.mn.org
               399 Beacon Ave.
               St. Paul, MN  55104-3527

Permission to use, copy, modify, and distribute this software and
its documentation for any purpose and without fee is hereby
granted, provided that the above copyright notice appear in all
copies and that both that copyright notice and this permission
notice appear in supporting documentation, and that the name of
Dale Schumacher not be used in advertising or publicity pertaining to
distribution of the software without specific, written prior
permission.  Dale Schumacher makes no representations about the
suitability of this software for any purpose.  It is provided "as
is" without express or implied warranty.