[KERNEL32] Also "fix" hack of RtlAcquirePrivilege in SetSystemTime(). Spotted by Thomas.
Note for syssetup/rtl/kernel32 devs:
Make it crystal clear that this is a hack and has to be deleted once RAP is fine.
Also note that for SetLocalTime() and 2nd stage, syssetup acquires the privilege itself
in SetSystemLocalTime() before calling SetLocalTime().
This will also have to be deleted once RAP is working fine. Otherwise it will keep failing:
it is not possible to acquire a privilege already held.