[REGEDIT] Fix tree-view's image-list handle leakage when application quits. Loosely based on a patch by Edijs Kolesnicovičs and Grégory Macário Harbs
NOTE: Always cleanup / destroy (or, try to) things in the reverse way they are created / initialized (i.e. in a symmetrical way). Therefore, destry the associated tree-view's image-list in the DestroyTreeView function, which is the opposite of CreateTreeView (which calls InitTreeViewImageLists). The same mechanism is already used by the list-view. For completeness, add a parameter to the DestroyTreeView function (a handle to a tree-view) so that we can pass to it the global tree-view's handle (see what's done in WM_DESTROY message handling in ChildWndProc).

CORE-6856 #resolve #comment Should be fixed in revision r59371. See the commit log for more details. Thanks :)